+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/wallet/lib/data/rfc/tokenStatusList/MediaTypes.kt
@@ -9,7 +9,8 @@ data object MediaTypes {
 
         /** `application/statuslist+cwt` */
         const val STATUSLIST_CWT = "application/statuslist+cwt"
-
+        
+        /** `application/identifierlist+cwt` */
         const val IDENTIFIERLIST_CWT = "application/identifierlist+cwt"
 
         @Deprecated(