+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/InMemoryIssuerCredentialStore.kt
@@ -12,13 +12,13 @@ import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatusBitSize
 import com.benasher44.uuid.uuid4
 import kotlinx.coroutines.sync.Mutex
-import kotlin.collections.mutableListOf
 import kotlin.time.Instant
 
 class InMemoryIssuerCredentialStore(
     val tokenStatusBitSize: TokenStatusBitSize = TokenStatusBitSize.ONE,
 ) : IssuerCredentialStore, ReferencedTokenStore {
     private val indexMutex = Mutex()
+
     data class Credential(
         val vcId: String,
         val statusListIndex: ULong,
@@ -31,7 +31,15 @@ class InMemoryIssuerCredentialStore(
     private val credentialMap = mutableMapOf<Int, MutableList<Credential>>()
 
     /** Tracks revoked identifiers for timePeriod to build [IdentifierList]; Sets to remove duplicates */
-    private val identifierRevokationList = mutableMapOf<Int, MutableSet<String>>()
+    private val identifierRevocationList = mutableMapOf<Int, MutableSet<String>>()
+
+    @Deprecated("Renamed", replaceWith = ReplaceWith("createStoredCredentialReference"))
+    override suspend fun createStatusListIndex(
+        credential: CredentialToBeIssued,
+        timePeriod: Int
+    ): KmmResult<IssuerCredentialStore.StoredCredentialReference> =
+        createStoredCredentialReference(credential, timePeriod)
+
 
     override suspend fun createStoredCredentialReference(
         credential: CredentialToBeIssued,
@@ -88,10 +96,18 @@ class InMemoryIssuerCredentialStore(
     }
 
     override fun getRawIdentifierList(timePeriod: Int): Map<Identifier, IdentifierInfo> =
-        identifierRevokationList.getOrElse(timePeriod) { emptySet() }.associate {
+        identifierRevocationList.getOrElse(timePeriod) { emptySet() }.associate {
             Identifier(it.encodeToByteArray()) to IdentifierInfo()
         }
 
+    /**
+     * Set the [status] of the referenced token with this [index] for the [timePeriod], if it exists.
+     *
+     * If [status] is [TokenStatus.Invalid] the associated identifier will be added to [identifierRevocationList]
+     * Note that ISO 18-013 does not support any action besides full revocation. If a credential has been suspended it remains suspended.
+     *
+     * Care must be taken to handle drift between the two systems and it is recommended to use only one at a time.
+     */
     override fun setStatus(
         timePeriod: Int,
         index: ULong,
@@ -105,11 +121,17 @@ class InMemoryIssuerCredentialStore(
 
         entry.status = status
         if (status == TokenStatus.Invalid) {
-            identifierRevokationList.getOrPut(timePeriod) { mutableSetOf() }.add(entry.vcId)
+            identifierRevocationList.getOrPut(timePeriod) { mutableSetOf() }.add(entry.vcId)
         }
         return true
     }
 
+    /**
+     * Set the status of the referenced token with this [identifier] for the [timePeriod] to revoked, if it exists.
+     * Additionally the `TokenStatus` at the associated `StatusListIndex` is also automatically set to invalid
+     *
+     * ISO 18-013 does not support any action besides full revocation. If a credential has been suspended it remains suspended
+     */
     override fun revokeIdentifier(
         timePeriod: Int,
         identifier: ByteArray
@@ -120,7 +142,7 @@ class InMemoryIssuerCredentialStore(
             it.vcId == identifier.decodeToString()
         } ?: return false
 
-        identifierRevokationList.getOrPut(timePeriod) { mutableSetOf() }.add(entry.vcId)
+        identifierRevocationList.getOrPut(timePeriod) { mutableSetOf() }.add(entry.vcId)
         entry.status = TokenStatus.Invalid
         return true
     }
@@ -128,7 +150,9 @@ class InMemoryIssuerCredentialStore(
 
 private val Issuer.IssuedCredential.validUntil: Instant
     get() = when (this) {
-        is Issuer.IssuedCredential.Iso -> this.issuerSigned.issuerAuth.payload?.validityInfo?.validUntil ?: Instant.DISTANT_PAST
-        is Issuer.IssuedCredential.VcJwt -> this.vc.expirationDate?: Instant.DISTANT_PAST
+        is Issuer.IssuedCredential.Iso -> this.issuerSigned.issuerAuth.payload?.validityInfo?.validUntil
+            ?: Instant.DISTANT_PAST
+
+        is Issuer.IssuedCredential.VcJwt -> this.vc.expirationDate ?: Instant.DISTANT_PAST
         is Issuer.IssuedCredential.VcSdJwt -> this.sdJwtVc.expiration ?: Instant.DISTANT_PAST
     }