+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Release 5.12.0 (unreleased):
    - Change: Executing unsatisfiable DCQL queries no longer throws on matching, only on submission.
    - Change: Update DCQLClaimsQuery to OpenID4VP 1.0
    - Change: Do not fail when only matching credentials without submitting a presentation
+   - Allow issuance and verification of `IdentifierList` Revocation Mechanism
  - OpenID for Verifiable Credential Issuance:
    - Moved the class `RefreshTokenInfo` from `OpenId4VciClient` to `SubjectCredentialStore.kt` and renamed it to `CredentialRenewalInfo` to better describe its role in the renewal process.
      Kept `RefreshTokenInfo` in the original package for backward compatibility